update entities and helpers
